回购加4元

+++ app/src/main/java/com/studentloan/white/net/data/JieKuanFeiYong.java
@@ -16,7 +16,7 @@ public class JieKuanFeiYong {
 
     public int jinE;
     public float feiYong;
-    public int daoZhangJinE;
+    public float daoZhangJinE;
     public int tianShu;
     public float liXi;
     public float yingHuanJinE;
